requirements.txt
langchain
openai
python-dotenv
huggingface_hub
cassio
datasets
tiktoken
PyPDF2
sentence-transformers
llama-index-llms-huggingface
transformers
einops
accelerate
pypdf
llama-index
torch